# api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/stations": {
      "post": {
        "operationId": "StationsController_create",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a station in the current tenant.",
        "tags": [
          "Stations"
        ]
      },
      "get": {
        "operationId": "StationsController_list",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "central",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedStationsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List stations inside the current tenant.",
        "tags": [
          "Stations"
        ]
      }
    },
    "/stations/{id}": {
      "get": {
        "operationId": "StationsController_getById",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Station not found in current tenant."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a station detail by id in the current tenant.",
        "tags": [
          "Stations"
        ]
      },
      "patch": {
        "operationId": "StationsController_update",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Station not found in current tenant."
          },
          "409": {
            "description": "Deactivating the station would leave an active line using it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update station fields in the current tenant.",
        "tags": [
          "Stations"
        ]
      },
      "put": {
        "operationId": "StationsController_replace",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Station not found in current tenant."
          },
          "409": {
            "description": "Deactivating the station would leave an active line using it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Replace station fields in the current tenant.",
        "tags": [
          "Stations"
        ]
      },
      "delete": {
        "operationId": "StationsController_remove",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Station not found in current tenant."
          },
          "409": {
            "description": "Station is referenced by lines or reservations."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a station in the current tenant if no references exist.",
        "tags": [
          "Stations"
        ]
      }
    },
...
  },
  "info": {
    "title": "SURP API",
    "description": "Backend API for SURP, providing tenant-scoped authentication, operational health endpoints, and secure session management for transportation workflows. Protected endpoints require BOTH Authorization: Bearer <accessToken> and X-Tenant-Slug headers (except /platform/* routes), and the tenant must match the token claim.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token issued by /auth/login. Use together with X-Tenant-Slug on protected endpoints."
      }
    },
    "schemas": {
...
      "CreateStationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Central Station"
          },
          "address": {
            "type": "string",
            "example": "1 Main Street, Demo City"
          },
          "category": {
            "type": "string",
            "enum": [
              "BUS_STATION",
              "BUS_STOP"
            ]
          },
          "contactPhone": {
            "type": "string",
            "example": "+1-555-0100"
          },
          "notes": {
            "type": "string",
            "example": "Primary station near downtown terminal."
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "default": true
          }
        },
        "required": [
          "name",
          "address"
        ]
      },
      "StationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "createdById": {
            "type": "object",
            "nullable": true
          },
          "updatedById": {
            "type": "object",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "BUS_STATION",
              "BUS_STOP"
            ],
            "nullable": true
          },
          "contactPhone": {
            "type": "object",
            "nullable": true
          },
          "notes": {
            "type": "object",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "createdById",
          "updatedById",
          "name",
          "address",
          "category",
          "contactPhone",
          "notes",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedStationsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StationResponseDto"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 25
          },
          "total": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "UpdateStationDto": {
        "type": "object",
        "properties": {
          "confirmBreakingChange": {
            "type": "boolean",
            "example": false,
            "description": "Confirms a change previously refused with WOULD_BREAK_RESERVATIONS."
          },
          "name": {
            "type": "string",
            "example": "Central Station - West Gate"
          },
          "address": {
            "type": "string",
            "example": "10 West Street, Demo City"
          },
          "category": {
            "type": "string",
            "enum": [
              "BUS_STATION",
              "BUS_STOP"
            ]
          },
          "contactPhone": {
            "type": "string",
            "example": "+1-555-0109"
          },
          "notes": {
            "type": "string",
            "example": "Updated boarding area information."
          },
          "isActive": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "WouldBreakReservationsDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "WOULD_BREAK_RESERVATIONS"
          },
          "invariant": {
            "type": "string",
            "example": "reservation.seatWithinCapacity",
            "description": "Key of the invariant the change would violate."
          },
          "affectedCount": {
            "type": "number",
            "example": 8,
            "description": "New invariant violations the proposed change would introduce."
          },
          "message": {
            "type": "string",
            "example": "Ova izmena cini 8 rezervacija nevidljivim."
          }
        },
        "required": [
          "code",
          "invariant",
          "affectedCount",
          "message"
        ]
      },
...
    }
  }
}